cleanup input & price indication
