+++ src/components/simulator/input/SimInputRecurring.tsx
@@ -56,7 +56,6 @@ export const SimInputRecurring = ({
         isOrdersReversed={checkIfOrdersReversed(state.buy, state.sell)}
         strategyType="recurring"
         isBudgetOptional={+state.sell.budget === 0 && +state.buy.budget > 0}
-        ignoreMarketPriceWarning
         warningMsg={warningMsg.sell}
       />
 
@@ -70,7 +69,6 @@ export const SimInputRecurring = ({
         isOrdersReversed={checkIfOrdersReversed(state.buy, state.sell)}
         strategyType="recurring"
         isBudgetOptional={+state.buy.budget === 0 && +state.sell.budget > 0}
-        ignoreMarketPriceWarning
         warningMsg={warningMsg.buy}
       />
     </>
